move: ResultCalculator 패키지 이동
