+++ src/main/kotlin/lotto/service/ResultCalculator.kt
@@ -1,6 +1,9 @@
-package lotto.model
+package lotto.service
 
 import lotto.constants.LottoPrize
+import lotto.model.LottoStore
+import lotto.model.PurchaseInfo
+import lotto.model.WinningLotto
 import kotlin.math.floor
 
 object ResultCalculator {
